Full implementation of the tag model

+++ src/main/java/com/dosvales/vagoapp/dao/TagDao.java
@@ -15,6 +15,6 @@ public interface TagDao extends GenericDao<Tag, Long> {
 
 	Integer existenceOfTags();
 
-	Long findLastConsecutive(TypeMovement typeMovement);
+	Integer findLastConsecutive(TypeMovement typeMovement);
 
 }
